+++ tests/testthat/test-add.R
@@ -55,6 +55,64 @@ describe(
 )
 
 describe(
+  "add_route()", {
+    ppath = with_temp_package()
+    it("will error when package is not a pipework project", {
+      expect_error(add_route())
+    })
+    create_lock_file()
+    it("will add a simple route /test when no routes exist", {
+      expect_false(file.exists("inst/extdata/api/routes/_.R"))
+      expect_false(file.exists("R/api_.R"))
+      expect_false(file.exists("R/test.R"))
+      add_route("test")
+      # creates 3 new files
+      expect_true(file.exists("inst/extdata/api/routes/_.R"))
+      expect_true(file.exists("R/api_.R"))
+      expect_true(file.exists("R/test.R"))
+      # make sure the files are parsable
+      expect_type(parse("inst/extdata/api/routes/_.R"), "expression")
+      expect_type(parse("R/api_.R"), "expression")
+      expect_type(parse("R/test.R"), "expression")
+    })
+    it("will add a simple route /test2 appending to files", {
+      expect_true(file.exists("inst/extdata/api/routes/_.R"))
+      base_route_lc = length(readLines("inst/extdata/api/routes/_.R"))
+      expect_true(file.exists("R/api_.R"))
+      api_lc = length(readLines("R/api_.R"))
+      expect_false(file.exists("R/test2.R"))
+      add_route("test2")
+      # creates 1 new file
+      expect_true(file.exists("R/test2.R"))
+      # existing files are appended to
+      expect_gt(length(readLines("inst/extdata/api/routes/_.R")), base_route_lc)
+      expect_gt(length(readLines("R/api_.R")), api_lc)
+    })
+    it("will add a file route test/is_alive", {
+      add_route("test/is_alive")
+      # creates 3 new files
+      expect_true(file.exists("inst/extdata/api/routes/test.R"))
+      expect_true(file.exists("R/api_test.R"))
+      expect_true(file.exists("R/is_alive.R"))
+    })
+    it("will add a nested file route nested/test/is_alive", {
+      add_route("nested/test/is_alive")
+      # creates 3 new files
+      expect_true(file.exists("inst/extdata/api/routes/nested/test.R"))
+      expect_true(file.exists("R/api_nested_test.R"))
+      expect_true(file.exists("R/is_alive.R"))
+    })
+    it("will add the plumber route file if it doesn't exist for a deep nested file route", {
+      add_route("nested/twice/test/is_alive")
+      # creates 3 new files
+      expect_true(file.exists("inst/extdata/api/routes/nested/twice/test.R"))
+      expect_true(file.exists("R/api_nested_twice_test.R"))
+      expect_true(file.exists("R/is_alive.R"))
+    })
+  }
+)
+
+describe(
   "add_hooks()", {
     it("will error when package is not a pipework project", {
       ppath = with_temp_package()